Automate opening of Microsoft Teams
I tried several of different ways to pipe things into docker:
 * Using `yes`
 * Using `printf`
 * Using a dozen variations of a hand-written C program

Only the `yes` command works. I literally created a mock-up of the `yes` command in C and it didn't work. `yes` is doing some magic under the hood and I don't know what this magic is, so let's just use `yes`. Yes, `yes` is evil when misused in this way, and I really wish there were an alternative.